refactor(customer): move the first_user_event_at backfill query to the script
