+++ server/tests/integrations/tinybird/test_service.py
@@ -2,12 +2,10 @@ import asyncio
 import json
 import uuid
 from datetime import UTC, datetime, timedelta
-from unittest.mock import AsyncMock
 
 import httpx
 import pytest
 import respx
-from pytest_mock import MockerFixture
 
 from polar.integrations.tinybird.client import (
     MAX_RETRIES,
@@ -21,7 +19,6 @@ from polar.integrations.tinybird.service import (
     TinybirdEventTypesQuery,
     _event_to_tinybird,
     count_user_events_by_organization,
-    get_first_user_event_at_by_organization,
 )
 from polar.models import Event
 from polar.models.event import EventSource
@@ -674,55 +671,3 @@ class TestTinybirdRequestError:
 
             with pytest.raises(TinybirdOperationalError, match="500"):
                 await client.endpoint("metrics")
-
-
-@pytest.mark.asyncio
-class TestGetFirstUserEventAtByOrganization:
-    async def test_pages_until_a_short_page(self, mocker: MockerFixture) -> None:
-        mocker.patch("polar.integrations.tinybird.service.FIRST_SEEN_PAGE_SIZE", 2)
-        customer_ids = [uuid.uuid4() for _ in range(3)]
-        timestamp = datetime(2020, 1, 1, tzinfo=UTC)
-        query_mock = mocker.patch(
-            "polar.integrations.tinybird.service.client.query",
-            new_callable=AsyncMock,
-            side_effect=[
-                [
-                    {"customer_id": str(id), "first_seen": timestamp}
-                    for id in customer_ids[:2]
-                ],
-                [{"customer_id": str(customer_ids[2]), "first_seen": timestamp}],
-                [],
-            ],
-        )
-
-        (
-            by_customer_id,
-            by_external_customer_id,
-        ) = await get_first_user_event_at_by_organization(uuid.uuid4())
-
-        assert set(by_customer_id) == set(customer_ids)
-        assert by_external_customer_id == {}
-        # Two pages for the customer id view, one empty for the external one.
-        assert query_mock.await_count == 3
-        # The second page resumes after the last key of the first, cast so
-        # ClickHouse compares UUIDs rather than a UUID against a string.
-        assert (
-            f"customer_id` > toUUID('{customer_ids[1]}')"
-            in query_mock.await_args_list[1].args[0]
-        )
-
-    async def test_no_rows(self, mocker: MockerFixture) -> None:
-        query_mock = mocker.patch(
-            "polar.integrations.tinybird.service.client.query",
-            new_callable=AsyncMock,
-            return_value=[],
-        )
-
-        (
-            by_customer_id,
-            by_external_customer_id,
-        ) = await get_first_user_event_at_by_organization(uuid.uuid4())
-
-        assert by_customer_id == {}
-        assert by_external_customer_id == {}
-        assert query_mock.await_count == 2
