+++ server/tests/scripts/test_backfill_customer_first_user_event_at.py
@@ -1,4 +1,6 @@
+import uuid
 from datetime import UTC, datetime
+from unittest.mock import AsyncMock
 
 import pytest
 from pytest_mock import MockerFixture
@@ -6,7 +8,10 @@ from sqlalchemy import select
 
 from polar.kit.db.postgres import AsyncSession
 from polar.models import Customer, Organization
-from scripts.backfill_customer_first_user_event_at import backfill_organization
+from scripts.backfill_customer_first_user_event_at import (
+    backfill_organization,
+    get_first_user_event_at_by_organization,
+)
 from tests.fixtures.database import SaveFixture
 from tests.fixtures.random_objects import create_customer
 
@@ -27,7 +32,7 @@ def _mock_views(
 ) -> None:
     mocker.patch(
         "scripts.backfill_customer_first_user_event_at."
-        "tinybird_service.get_first_user_event_at_by_organization",
+        "get_first_user_event_at_by_organization",
         return_value=(by_customer_id, by_external_customer_id),
     )
 
@@ -116,3 +121,55 @@ class TestBackfillOrganization:
 
         assert count == 1
         assert await _get_first_user_event_at(session, customer) is None
+
+
+@pytest.mark.asyncio
+class TestGetFirstUserEventAtByOrganization:
+    async def test_pages_until_a_short_page(self, mocker: MockerFixture) -> None:
+        mocker.patch("scripts.backfill_customer_first_user_event_at.PAGE_SIZE", 2)
+        customer_ids = [uuid.uuid4() for _ in range(3)]
+        timestamp = datetime(2020, 1, 1, tzinfo=UTC)
+        query_mock = mocker.patch(
+            "scripts.backfill_customer_first_user_event_at.tinybird_client.query",
+            new_callable=AsyncMock,
+            side_effect=[
+                [
+                    {"customer_id": str(id), "first_seen": timestamp}
+                    for id in customer_ids[:2]
+                ],
+                [{"customer_id": str(customer_ids[2]), "first_seen": timestamp}],
+                [],
+            ],
+        )
+
+        (
+            by_customer_id,
+            by_external_customer_id,
+        ) = await get_first_user_event_at_by_organization(uuid.uuid4())
+
+        assert set(by_customer_id) == set(customer_ids)
+        assert by_external_customer_id == {}
+        # Two pages for the customer id view, one empty for the external one.
+        assert query_mock.await_count == 3
+        # The second page resumes after the last key of the first, cast so
+        # ClickHouse compares UUIDs rather than a UUID against a string.
+        assert (
+            f"customer_id` > toUUID('{customer_ids[1]}')"
+            in query_mock.await_args_list[1].args[0]
+        )
+
+    async def test_no_rows(self, mocker: MockerFixture) -> None:
+        query_mock = mocker.patch(
+            "scripts.backfill_customer_first_user_event_at.tinybird_client.query",
+            new_callable=AsyncMock,
+            return_value=[],
+        )
+
+        (
+            by_customer_id,
+            by_external_customer_id,
+        ) = await get_first_user_event_at_by_organization(uuid.uuid4())
+
+        assert by_customer_id == {}
+        assert by_external_customer_id == {}
+        assert query_mock.await_count == 2
